Validate cause area input
